Make output order of resources like input order
To preserve the CSV order of &resources one has to manually insert &sortby='FIELD(modResource.id,[[+myResources]])'. This is a possible way how do do it but I thought it would be nicer if we had the sortby token resources which ads this behavior automatically.

I hope you like my little addition and will merge it. Or just feel free to modify it.

cheers
Oliver